Refactoring & Restructing for simplicity & Improved reusability from integration tests
