+++ t/smoke/conftest.py
@@ -1,10 +1,94 @@
-from typing import Tuple
+import os
+from typing import Tuple, Type
 
 import pytest
+from pytest_celery import defaults
 from pytest_celery.api.components.worker.cluster import CeleryWorkerCluster
 from pytest_celery.api.components.worker.node import CeleryTestWorker
+from pytest_celery.containers.worker import CeleryWorkerContainer
+from pytest_docker_tools import build, container, fxtr
 
-from t.smoke.workers import *  # noqa
+from celery import Celery
+from t.smoke.workers import Celery4WorkerContainer, SmokeWorkerContainer
+
+# Celery 4.x
+
+celery4_worker_image = build(
+    path=".",
+    dockerfile="t/smoke/workers/4.4.7",
+    tag="t/smoke/worker:celery4",
+    buildargs=Celery4WorkerContainer.buildargs(),
+)
+
+
+celery4_worker_container = container(
+    image="{celery4_worker_image.id}",
+    environment=fxtr("default_worker_env"),
+    network="{DEFAULT_NETWORK.name}",
+    volumes={"{default_worker_volume.name}": defaults.DEFAULT_WORKER_VOLUME},
+    wrapper_class=Celery4WorkerContainer,
+    timeout=defaults.DEFAULT_WORKER_CONTAINER_TIMEOUT,
+)
+
+
+@pytest.fixture
+def celery4_worker(celery4_worker_container: CeleryWorkerContainer, celery_setup_app: Celery) -> CeleryTestWorker:
+    yield CeleryTestWorker(celery4_worker_container, app=celery_setup_app)
+
+
+# Dev worker
+
+smoke_worker_image = build(
+    path=".",
+    dockerfile="t/smoke/workers/dev",
+    tag="t/smoke/worker:dev",
+    buildargs=SmokeWorkerContainer.buildargs(),
+)
+
+
+default_worker_container = container(
+    image="{smoke_worker_image.id}",
+    environment=fxtr("default_worker_env"),
+    network="{DEFAULT_NETWORK.name}",
+    volumes={
+        # Volume: Worker /app
+        "{default_worker_volume.name}": defaults.DEFAULT_WORKER_VOLUME,
+        # Mount: Celery source
+        os.path.abspath(os.getcwd()): {
+            "bind": "/celery",
+            "mode": "rw",
+        },
+    },
+    wrapper_class=SmokeWorkerContainer,
+    timeout=defaults.DEFAULT_WORKER_CONTAINER_TIMEOUT,
+)
+
+
+@pytest.fixture
+def default_worker_container_cls() -> Type[CeleryWorkerContainer]:
+    return SmokeWorkerContainer
+
+
+@pytest.fixture(scope="session")
+def default_worker_container_session_cls() -> Type[CeleryWorkerContainer]:
+    return SmokeWorkerContainer
+
+
+# Tasks
+
+
+@pytest.fixture
+def default_worker_tasks() -> set:
+    from t.integration import tasks as integration_tests_tasks
+    from t.smoke import tasks as smoke_tests_tasks
+
+    yield {
+        integration_tests_tasks,
+        smoke_tests_tasks,
+    }
+
+
+# Workers cluster
 
 
 @pytest.fixture(
