--- t/smoke/workers/celery4/__init__.py
@@ -1,5 +1,0 @@
-# flake8: noqa
-
-
-from t.smoke.workers.celery4.api import Celery4WorkerContainer
-from t.smoke.workers.celery4.fixtures import *  # noqa
