+++ t/smoke/workers/dev
@@ -9,7 +9,7 @@ RUN apt-get update && apt-get install -y build-essential
 # Set arguments
 ARG CELERY_VERSION
 ARG CELERY_LOG_LEVEL=INFO
-ARG CELERY_WORKER_NAME=celery_test_worker
+ARG CELERY_WORKER_NAME=celery_dev_worker
 ARG CELERY_WORKER_QUEUE=celery
 ENV PIP_VERSION=$CELERY_VERSION
 ENV LOG_LEVEL=$CELERY_LOG_LEVEL
