--- t/smoke/workers/dev/fixtures.py
@@ -1,55 +1,0 @@
-import os
-from typing import Type
-
-import pytest
-from pytest_celery import defaults
-from pytest_celery.containers.worker import CeleryWorkerContainer
-from pytest_docker_tools import build, container, fxtr
-
-from t.smoke.workers.dev.api import SmokeWorkerContainer
-
-smoke_worker_image = build(
-    path=".",
-    dockerfile="t/smoke/workers/dev/Dockerfile",
-    tag="t/smoke/worker:dev",
-    buildargs=SmokeWorkerContainer.buildargs(),
-)
-
-
-default_worker_container = container(
-    image="{smoke_worker_image.id}",
-    environment=fxtr("default_worker_env"),
-    network="{DEFAULT_NETWORK.name}",
-    volumes={
-        # Volume: Worker /app
-        "{default_worker_volume.name}": defaults.DEFAULT_WORKER_VOLUME,
-        # Mount: Celery source
-        os.path.abspath(os.getcwd()): {
-            "bind": "/celery",
-            "mode": "rw",
-        },
-    },
-    wrapper_class=SmokeWorkerContainer,
-    timeout=defaults.DEFAULT_WORKER_CONTAINER_TIMEOUT,
-)
-
-
-@pytest.fixture
-def default_worker_container_cls() -> Type[CeleryWorkerContainer]:
-    return SmokeWorkerContainer
-
-
-@pytest.fixture(scope="session")
-def default_worker_container_session_cls() -> Type[CeleryWorkerContainer]:
-    return SmokeWorkerContainer
-
-
-@pytest.fixture
-def default_worker_tasks() -> set:
-    from t.integration import tasks as integration_tests_tasks
-    from t.smoke import tasks as smoke_tests_tasks
-
-    yield {
-        integration_tests_tasks,
-        smoke_tests_tasks,
-    }
