+++ t/smoke/workers/latest
@@ -1,4 +1,4 @@
-FROM python:3.10-slim-buster
+FROM python:3.11-slim-buster
 
 # Create a user to run the worker
 RUN adduser --disabled-password --gecos "" test_user
@@ -7,9 +7,9 @@ RUN adduser --disabled-password --gecos "" test_user
 RUN apt-get update && apt-get install -y build-essential
 
 # Set arguments
-ARG CELERY_VERSION="4.4.7"
+ARG CELERY_VERSION
 ARG CELERY_LOG_LEVEL=INFO
-ARG CELERY_WORKER_NAME=celery4_worker
+ARG CELERY_WORKER_NAME=celery_latest_worker
 ARG CELERY_WORKER_QUEUE=celery
 ENV PIP_VERSION=$CELERY_VERSION
 ENV LOG_LEVEL=$CELERY_LOG_LEVEL
@@ -21,7 +21,7 @@ ENV PYTHONDONTWRITEBYTECODE 1
 
 # Install Python dependencies
 RUN pip install --no-cache-dir --upgrade pip \
-    && pip install --no-cache-dir celery[redis]==$PIP_VERSION
+    && pip install --no-cache-dir celery[redis]
 
 # The workdir must be /app
 WORKDIR /app
